requirements.txt
clean-fid
colab-convert
einops
ffmpeg
ftfy
ipython
jsonmerge
jupyterlab
jupyter_http_over_ws
kornia
matplotlib
notebook
numexpr
omegaconf
opencv-python
pandas
pytorch_lightning
resize-right
scikit-image
scikit-learn
timm
torchdiffeq
transformers==4.19.2
